+++ types/Fee/FeeType.d.ts
@@ -1,0 +1,1 @@
+export declare type TFeeType = 'LabelFee' | 'PostageFee' | 'InsuranceFee' | 'TrackerFee';
