+++ types/Fee/index.d.ts
@@ -1,0 +1,2 @@
+export * from './Fee';
+export * from './FeeType';
