+++ types/Tracker/Carrier.d.ts
@@ -1,0 +1,85 @@
+/**
+ * @see https://www.easypost.com/docs/api/node#carrier-tracking-strings
+ */
+export declare type Carrier =
+  | 'AmazonMws'
+  | 'APC'
+  | 'Aramex'
+  | 'ArrowXL'
+  | 'Asendia'
+  | 'AustraliaPost'
+  | 'AxlehireV3'
+  | 'BorderGuru'
+  | 'Cainiao'
+  | 'CanadaPost'
+  | 'Canpar'
+  | 'ColumbusLastMile'
+  | 'Chronopost'
+  | 'ColisPrive'
+  | 'Colissimo'
+  | 'CouriersPlease'
+  | 'DaiPost'
+  | 'Deliv'
+  | 'DeutschePost'
+  | 'DeutschePostUK'
+  | 'DHLEcommerceAsia'
+  | 'DhlEcs'
+  | 'DHLExpress'
+  | 'DHLFreight'
+  | 'DHLGermany'
+  | 'Dicom'
+  | 'DPD'
+  | 'DPDUK'
+  | 'ChinaEMS'
+  | 'Estafeta'
+  | 'Estes'
+  | 'Fastway'
+  | 'FedEx'
+  | 'FedExCrossBorder'
+  | 'FedExMailview'
+  | 'FedExSameDayCity'
+  | 'FedexSmartPost'
+  | 'FirstMile'
+  | 'Globegistics'
+  | 'GSO'
+  | 'Hermes'
+  | 'HongKongPost'
+  | 'InterlinkExpress'
+  | 'JancoFreight'
+  | 'JPPost'
+  | 'KuronekoYamato'
+  | 'LaPoste'
+  | 'LaserShipV2'
+  | 'Liefery'
+  | 'LoomisExpress'
+  | 'LSO'
+  | 'Newgistics'
+  | 'OnTrac'
+  | 'OnTracDirectPost'
+  | 'OsmWorldwide'
+  | 'Parcelforce'
+  | 'Passport'
+  | 'Pilot'
+  | 'PostNL'
+  | 'Posten'
+  | 'PostNord'
+  | 'Purolator'
+  | 'RoyalMail'
+  | 'RRDonnelley'
+  | 'Seko'
+  | 'OmniParcel'
+  | 'SingaporePost'
+  | 'SpeeDee'
+  | 'SprintShip'
+  | 'StarTrack'
+  | 'Toll'
+  | 'TForce'
+  | 'UDS'
+  | 'Ukrposhta'
+  | 'UPS'
+  | 'UPSIparcel'
+  | 'UPSMailInnovations'
+  | 'USPS'
+  | 'Veho'
+  | 'Yanwen'
+  | 'Yodel';
