+++ types/Tracker/TrackerStatus.d.ts
@@ -1,0 +1,11 @@
+export declare type TTrackerStatus =
+  | 'unknown'
+  | 'pre_transit'
+  | 'in_transit'
+  | 'out_for_delivery'
+  | 'delivered'
+  | 'available_for_pickup'
+  | 'return_to_sender'
+  | 'failure'
+  | 'cancelled'
+  | 'error';
