+++ types/Tracker/index.d.ts
@@ -1,0 +1,8 @@
+export * from './Carrier';
+export * from './CarrierDetail';
+export * from './Tracker';
+export * from './TrackerStatus';
+export * from './TrackingDetail';
+export * from './TrackingLocation';
+export * from './TrackingStatus';
+export * from './TrackerCreateParameters';
