+++ types/index.d.ts
@@ -1,94 +1,10 @@
 import { Address, IAddress, IAddressCreateParameters, IVerifications } from './Address';
 import { IBaseObject, IDatedObject, IObjectWithId } from './base';
+import { IFee } from './Fee';
 import { PredefinedPackage } from './PredefinedPackage';
 import { ServiceLevel } from './ServiceLevel';
-import { DeepPartial, ParametersToOmitOnCreate } from './utils';
-
-/**
- * @see https://www.easypost.com/docs/api/node#carrier-tracking-strings
- */
-export declare type Carrier =
-  | 'AmazonMws'
-  | 'APC'
-  | 'Aramex'
-  | 'ArrowXL'
-  | 'Asendia'
-  | 'AustraliaPost'
-  | 'AxlehireV3'
-  | 'BorderGuru'
-  | 'Cainiao'
-  | 'CanadaPost'
-  | 'Canpar'
-  | 'ColumbusLastMile'
-  | 'Chronopost'
-  | 'ColisPrive'
-  | 'Colissimo'
-  | 'CouriersPlease'
-  | 'DaiPost'
-  | 'Deliv'
-  | 'DeutschePost'
-  | 'DeutschePostUK'
-  | 'DHLEcommerceAsia'
-  | 'DhlEcs'
-  | 'DHLExpress'
-  | 'DHLFreight'
-  | 'DHLGermany'
-  | 'Dicom'
-  | 'DPD'
-  | 'DPDUK'
-  | 'ChinaEMS'
-  | 'Estafeta'
-  | 'Estes'
-  | 'Fastway'
-  | 'FedEx'
-  | 'FedExCrossBorder'
-  | 'FedExMailview'
-  | 'FedExSameDayCity'
-  | 'FedexSmartPost'
-  | 'FirstMile'
-  | 'Globegistics'
-  | 'GSO'
-  | 'Hermes'
-  | 'HongKongPost'
-  | 'InterlinkExpress'
-  | 'JancoFreight'
-  | 'JPPost'
-  | 'KuronekoYamato'
-  | 'LaPoste'
-  | 'LaserShipV2'
-  | 'Liefery'
-  | 'LoomisExpress'
-  | 'LSO'
-  | 'Newgistics'
-  | 'OnTrac'
-  | 'OnTracDirectPost'
-  | 'OsmWorldwide'
-  | 'Parcelforce'
-  | 'Passport'
-  | 'Pilot'
-  | 'PostNL'
-  | 'Posten'
-  | 'PostNord'
-  | 'Purolator'
-  | 'RoyalMail'
-  | 'RRDonnelley'
-  | 'Seko'
-  | 'OmniParcel'
-  | 'SingaporePost'
-  | 'SpeeDee'
-  | 'SprintShip'
-  | 'StarTrack'
-  | 'Toll'
-  | 'TForce'
-  | 'UDS'
-  | 'Ukrposhta'
-  | 'UPS'
-  | 'UPSIparcel'
-  | 'UPSMailInnovations'
-  | 'USPS'
-  | 'Veho'
-  | 'Yanwen'
-  | 'Yodel';
+import { Carrier, ITracker, Tracker } from './Tracker';
+import { DeepPartial, IAllMethodParameters, ParametersToOmitOnCreate } from './utils';
 
 /**
  * Parcel objects represent the physical container being shipped.
@@ -210,266 +126,6 @@ export declare interface IInsurance extends IObjectWithId<'Insurance'>, IDatedOb
    * The list of errors encountered during attempted purchase of the insurance
    */
   messages: string[];
-}
-
-export declare type TFeeType = 'LabelFee' | 'PostageFee' | 'InsuranceFee' | 'TrackerFee';
-
-/**
- * Fee objects are used to represent the breakdown of charges made when purchasing an item on EasyPost.
- * Shipments and Trackers both have associations to Fee objects.
- *
- * Each Shipment object will have a Fee of type "LabelFee" to represent the label fee charged by EasyPost for the service.
- * Shipments with postage collected by EasyPost (as opposed to shipments with postage collected directly by the carrier) will have a "PostageFee" according to the postage amount.
- * Insurance on a Shipment will add an "InsuranceFee" with the insurance premium (not the covered value) for the amount.
- * Tracker objects will have a "TrackerFee" with the price, even when a Tracker is free.
- *
- * @see https://www.easypost.com/docs/api/node#fee-object
- */
-export declare interface IFee extends IObjectWithId<'Fee'> {
-  /**
-   * The name of the category of fee. Possible types are "LabelFee", "PostageFee", "InsuranceFee", and "TrackerFee"
-   */
-  type: TFeeType;
-
-  /**
-   * USD value with sub-cent precision
-   */
-  amount: string;
-
-  /**
-   * Whether EasyPost has successfully charged your account for the fee
-   */
-  charged: boolean;
-
-  /**
-   * Whether the Fee has been refunded successfully
-   */
-  refunded: boolean;
-}
-
-export declare type TTrackerStatus =
-  | 'unknown'
-  | 'pre_transit'
-  | 'in_transit'
-  | 'out_for_delivery'
-  | 'delivered'
-  | 'available_for_pickup'
-  | 'return_to_sender'
-  | 'failure'
-  | 'cancelled'
-  | 'error';
-
-/**
- * A Tracker object contains all of the tracking information for a package.
- * A Tracker is created automatically whenever you buy a Shipment through EasyPost.
- * Ff you don't use EasyPost to purchase your shipping labels, you can still track packages through our API by creating a Tracker object directly.
- * Each Tracker is continually updated in the background as the package moves through its life cycle, regardless of whether or not the label was purchased through EasyPost.
- *
- * After creation, a Tracker object will be updated periodically based on when the carrier provides EasyPost with new tracking information.
- * This information can be consumed by using our webhooks infrastructure.
- * Every time a Tracker is updated a webhook Event will be sent.
- *
- * The Tracker object contains both the current information about the package as well as previous updates.
- * All of the previous updates are stored in the tracking_details array.
- * Each TrackingDetail object contains the status, the message from the carrier, and a TrackingLocation.
- *
- * The TrackingLocation contains city, state, country, and zip information about the location where the package was scanned.
- * The information each carrier provides is different, so some carriers may not make use of all of these fields.
- *
- * Some Tracker objects may also contain a CarrierDetail, which stores some additional information about the Tracker that the carrier has made available to EasyPost.
- * The CarrierDetail object contains the service and container_type of the package.
- * Additionally, it also stores the est_delivery_date_local and est_delivery_time_local, which provide information about the local delivery time.
- *
- * It's worth noting that tracking_codes are not globally unique.
- * Each carrier promises uniqueness for a given tracking_code for a certain period of time, but the length of time varies greatly based on the specific carrier and service level.
- * The carriers do eventually recycle tracking_codes, and for this reason enforcing uniqueness on the tracking_code field is not recommended.
- * EasyPost does, however, prevent the creation of duplicate Trackers based on tracking_code and carrier; duplicate requests by the same User will simply return the original Tracker.
- *
- * @see https://www.easypost.com/docs/api/node#tracker-object
- */
-export declare interface ITracker extends IObjectWithId<'Tracker'>, IDatedObject {
-  /**
-   * The tracking code provided by the carrier
-   */
-  tracking_code: string;
-
-  /**
-   * The current status of the package, possible values are "unknown", "pre_transit", "in_transit", "out_for_delivery", "delivered", "available_for_pickup", "return_to_sender", "failure", "cancelled" or "error"
-   */
-  status: TTrackerStatus;
-
-  /**
-   * The name of the person who signed for the package (if available)
-   */
-  signed_by: string;
-
-  /**
-   * The weight of the package as measured by the carrier in ounces (if available)
-   * float (oz)
-   */
-  weight: number;
-
-  /**
-   * The estimated delivery date provided by the carrier (if available)
-   */
-  est_delivery_date: string; // TODO maybe type should be Date?
-
-  /**
-   * The id of the EasyPost Shipment object associated with the Tracker (if any)
-   */
-  shipment_id: string;
-
-  /**
-   * The name of the carrier handling the shipment
-   */
-  carrier: string;
-
-  /**
-   * Array of the associated TrackingDetail objects
-   */
-  tracking_details: ITrackingDetail[];
-
-  /**
-   * The associated CarrierDetail object (if available)
-   */
-  carrier_detail: ICarrierDetail;
-
-  /**
-   * URL to a publicly-accessible html page that shows tracking details for this tracker
-   */
-  public_url: string;
-
-  /**
-   * Array of the associated Fee objects
-   */
-  fees: IFee[];
-}
-
-export declare type TTrackingStatus =
-  | 'unknown'
-  | 'pre_transit'
-  | 'in_transit'
-  | 'out_for_delivery'
-  | 'delivered'
-  | 'available_for_pickup'
-  | 'return_to_sender'
-  | 'failure'
-  | 'cancelled'
-  | 'error';
-
-/**
- * @see https://www.easypost.com/docs/api/node#tracking-detail-object
- */
-export declare interface ITrackingDetail extends IBaseObject<'TrackingDetail'> {
-  /**
-   * Description of the scan event
-   */
-  message: string;
-
-  /**
-   * Status of the package at the time of the scan event, possible values are "unknown", "pre_transit", "in_transit", "out_for_delivery", "delivered", "available_for_pickup", "return_to_sender", "failure", "cancelled" or "error"
-   */
-  status: TTrackingStatus;
-
-  /**
-   * The timestamp when the tracking scan occurred
-   */
-  datetime: string; // TODO maybe type should be Date?
-
-  /**
-   * The original source of the information for this scan event, usually the carrier
-   */
-  source: string;
-
-  /**
-   * The location associated with the scan event
-   */
-  tracking_location: ITrackingLocation;
-}
-
-/**
- * @see https://www.easypost.com/docs/api/node#tracking-location-object
- */
-export declare interface ITrackingLocation extends IBaseObject<'TrackingLocation'> {
-  /**
-   * The city where the scan event occurred (if available)
-   */
-  city: string;
-
-  /**
-   * The state where the scan event occurred (if available)
-   */
-  state: string;
-
-  /**
-   * The country where the scan event occurred (if available)
-   */
-  country: string;
-
-  /**
-   * The postal code where the scan event occurred (if available)
-   */
-  zip: string;
-}
-
-/**
- * @see https://www.easypost.com/docs/api/node#carrier-detail-object
- */
-export declare interface ICarrierDetail extends IBaseObject<'CarrierDetail'> {
-  /**
-   * The service level the associated shipment was shipped with (if available)
-   */
-  service: string;
-
-  /**
-   * The type of container the associated shipment was shipped in (if available)
-   */
-  container_type: string;
-
-  /**
-   * The estimated delivery date as provided by the carrier, in the local time zone (if available)
-   */
-  est_delivery_date_local: string;
-
-  /**
-   * The estimated delivery time as provided by the carrier, in the local time zone (if available)
-   */
-  est_delivery_time_local: string;
-
-  /**
-   * The location from which the package originated, stringified for presentation (if available)
-   */
-  origin_location: string;
-
-  /**
-   * The location from which the package originated, broken down by city/state/country/zip (if available)
-   */
-  origin_tracking_location: ITrackingLocation;
-
-  /**
-   * The location to which the package is being sent, stringified for presentation (if available)
-   */
-  destination_location: string;
-
-  /**
-   * The location to which the package is being sent, broken down by city/state/country/zip (if available)
-   */
-  destination_tracking_location: ITrackingLocation;
-
-  /**
-   * The date and time the carrier guarantees the package to be delivered by (if available)
-   */
-  guaranteed_delivery_date: string;
-
-  /**
-   * The alternate identifier for this package as provided by the carrier (if available)
-   */
-  alternate_identifier: string;
-
-  /**
-   * The date and time of the first attempt by the carrier to deliver the package (if available)
-   */
-  initial_delivery_attempt: string;
 }
 
 export declare type TBatchState =
@@ -1937,40 +1593,6 @@ export declare interface IUser extends IObjectWithId<'User'> {
   children: IUser[];
 }
 
-export declare interface IAllMethodParameters {
-  /**
-   * Optional pagination parameter.
-   * Only records created before the given ID will be included.
-   * May not be used with after_id
-   */
-  before_id?: string;
-
-  /**
-   * Optional pagination parameter.
-   * Only records created after the given ID will be included.
-   * May not be used with before_id
-   */
-  after_id?: string;
-
-  /**
-   * Only return records created after this timestamp.
-   * Defaults to 1 month ago, or 1 month before a passed end_datetime
-   */
-  start_datetime?: string;
-
-  /**
-   * Only return records created before this timestamp.
-   * Defaults to end of the current day, or 1 month after a passed start_datetime
-   */
-  end_datetime?: string;
-
-  /**
-   * The number of records to return on each page.
-   * The maximum value is 100
-   */
-  page_size?: number;
-}
-
 /**
  * @see https://www.easypost.com/docs/api/node#retrieve-a-list-of-shipments
  */
@@ -2373,91 +1995,6 @@ export declare class Insurance implements IInsurance {
    * @see https://www.easypost.com/docs/api/node#retrieve-an-insurance
    */
   static retrieve(insuranceId: string): Promise<Insurance>;
-}
-
-export declare interface ITrackerCreateParameters {
-  /**
-   * The tracking code associated with the package you'd like to track
-   */
-  tracking_code: string;
-
-  /**
-   * The carrier associated with the tracking_code you provided.
-   * The carrier will get auto-detected if none is provided
-   */
-  carrier?: Carrier;
-}
-
-export declare interface ITrackerListParameters extends IAllMethodParameters {
-  /**
-   * Only returns Trackers with the given tracking_code.
-   * Useful for retrieving an individual Tracker by tracking_code rather than by ID
-   */
-  tracking_code?: string;
-
-  /**
-   * Only returns Trackers with the given carrier value
-   */
-  carrier?: Carrier;
-}
-
-export declare class Tracker implements ITracker {
-  public constructor(input: ITrackerCreateParameters);
-
-  tracking_code: string;
-  status: TTrackerStatus;
-  signed_by: string;
-  weight: number;
-  est_delivery_date: string;
-  shipment_id: string;
-  carrier: string;
-  tracking_details: ITrackingDetail[];
-  carrier_detail: ICarrierDetail;
-  public_url: string;
-  fees: IFee[];
-  id: string;
-  mode: 'test' | 'production';
-  object: 'Tracker';
-  created_at: string;
-  updated_at: string;
-
-  /**
-   * A Tracker can be created with only a tracking_code.
-   * Optionally, you can provide the carrier parameter, which indicates the carrier the package was shipped with.
-   * If no carrier is provided, EasyPost will attempt to determine the carrier based on the tracking_code provided.
-   * Providing a carrier parameter is recommended, since some tracking_codes are ambiguous and may match with more than one carrier.
-   * In addition, not having to auto-match the carrier will significantly speed up the response time.
-   *
-   * In an effort to reduce wasted resources, EasyPost prevents the creation of duplicate Trackers.
-   * A Tracker is considered to be a duplicate if another Tracker with the same tracking_code and carrier was created by the same User in the last three months.
-   * In the case where a duplicate request is submitted, the original Tracker will be returned.
-   *
-   * @see https://www.easypost.com/docs/api/node#create-a-tracker
-   */
-  public save(): Promise<Tracker>;
-
-  /**
-   * The Tracker List is a paginated list of all Tracker objects associated with the given API Key.
-   * It accepts a variety of parameters which can be used to modify the scope.
-   * The has_more attribute indicates whether or not additional pages can be requested.
-   * The recommended way of paginating is to use either the before_id or after_id parameter to specify where the next page begins.
-   *
-   * Using the Tracker List endpoint is the recommended way of retrieving a Tracker by tracking_code or carrier.
-   * Unlike the retrieving a Tracker using the Retrieve endpoint, which accepts an id, the List endpoint accepts the tracking_code as the search parameter.
-   * Normally, you'll only have one Tracker with a given tracking_code, but it is also possible to further filter those results by including the carrier parameter in your request.
-   *
-   * @see https://www.easypost.com/docs/api/node#retrieve-a-list-of-trackers
-   */
-  static all(params?: ITrackerListParameters): Promise<{ trackers: Tracker[]; has_more: boolean }>;
-
-  /**
-   * Retrieve a Tracker by id.
-   *
-   * @param trackerId Unique, starts with "trk_"
-   *
-   * @see https://www.easypost.com/docs/api/node#retrieve-a-tracker
-   */
-  static retrieve(trackerId: string): Promise<Tracker>;
 }
 
 export declare interface IPickupCreateParameters {
